PIG-2128: Generating the jar file takes a lot of time and is unnecessary when running Pig local mode
git-svn-id: https://svn.apache.org/repos/asf/pig/trunk@1146491 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/apache/pig/backend/hadoop/executionengine/mapReduceLayer/JobControlCompiler.java
@@ -401,15 +401,16 @@ public class JobControlCompiler{
                 }
             }
 
-            if (!pigContext.inIllustrator) 
+            if (!pigContext.inIllustrator && pigContext.getExecType() != ExecType.LOCAL) 
             {
                 //Create the jar of all functions and classes required
                 File submitJarFile = File.createTempFile("Job", ".jar");
+                log.info("creating jar file "+submitJarFile.getName());
                 // ensure the job jar is deleted on exit
                 submitJarFile.deleteOnExit();
                 FileOutputStream fos = new FileOutputStream(submitJarFile);
                 JarManager.createJar(fos, mro.UDFs, pigContext);
-            
+                log.info("jar file "+submitJarFile.getName()+" created");
                 //Start setting the JobConf properties
                 conf.set("mapred.jar", submitJarFile.getPath());
             }
